20191011 1.fixed the warning of some cpp files 2.add js funtion in demo.py 3.add the dispatch funtion in ocx used by js 4.change the URL input dialog's default selection 5.add log file(Chinease) 6.update the readme file

+++ VideoActiveX/VideoActiveX/BlurThresDialog.cpp
@@ -63,6 +63,6 @@ void BlurThresDialog::OnBnClickedOk()
 {
 	// TODO: 在此添加控件通知处理程序代码
 	int thresindex = m_cb_blur_thres.GetCurSel();
-	threshold = (thresindex + 2) * 0.5;
+	threshold = static_cast<float>((thresindex + 2) * 0.5);
 	CDialogEx::OnOK();
 }
